Compute v2: Instance Networking Refactor
This commit does a large refactor of how networking is handled in
the openstack_compute_instance_v2 resource.

Two features were added:

1. The ability to attach an instance to the same network multiple
times.

2. The ability to use the Networking API and fall-back to the
legacy Nova Network API if required.

The network code was refactored and placed into a dedicated file
due to the number of functions required to determine an instance's
network configuration.

It is important to note that no changes were made to the Terraform
schema. Further, existing users should notice no difference in
the original functionality of the resource.