Fix Security Group setting on an OpenStack LBaaS
Using the `security_group_ids` parameter of an `openstack_lb_loadbalancer_v2` resource
leads to an `Invalid request due to incorrect syntax or missing required parameters.`
error message.

This is because terraform sends, in its request to OpenStack, an `allowed_address_pairs`
field with a value of `null` whereas OpenStack expects a list.

This change fixes the issue by not sending this field anymore.

Fixes #126