Networking v2: Fix allowed_address_pairs
This commit fixes a regression caused in #168.

This fix should now allow a user to both specify and omit a MAC
address when creating an allowed_address_pair and have the state
not require further changes. It should also allow importing of
ports with allowed_address_pairs to continue working.