Networking v2: Fix router distributed argument
This commit fixes the "distributed" argument for the
openstack_networking_router_v2 resource so that the value of
"false" is passed correctly.

"distributed" is removed from all acceptance tests since it is
now being passed correctly in the request body and this setting
is very dependent upon the testing environment configuration.