Return the default network uuid on instance create
Currently when an instance has no "network" blocks it's assumed it will
connect to the default network, worked out by inspecting the instance
addresses and taking the network name from that. This doesn't return the
network uuid, but the first refresh will pick the uuid up because the
resource now has a network block (with the default network name).

This commit simply copies some of the logic used in
getAllInstanceNetworks to work out network uuids from network names to
do the same with the default network name. Thus instances have a full
network block returned with name and uuid on create.