Networks v2: Network Data Source External Fix
PR #358 added support for querying networks by using the "external"
argument. However, this caused results to either contain only external
networks or only non-external networks.

When querying for an external network without the "external" argument,
no results would be found and Terraform would crash.

This fixes the behavior to only query with the "external" argument if
it is explicitly defined in the Terraform configuration.