Image Service v2: Make Properties non-computed
This commit changes the "properties" attribute to no longer be computed.
This is because OpenStack is adding an increasing amount of properties
to the image from the server-side which is causing conflicts with
user-generated properties.

A new attribute called "all_properties" has been added which will contain
all properties, both from the user and from the server.