Create an openstack_config data source
This data source loads an openstack config to memory in the format
expected by the provider. The use case for this is reading, creating
and uploading configs to nodes that need to access a specific OpenStack
cloud.

The data source takes as input the same schema used for the OpenStack
provider and generates a map with all the required configs to
authenticate to an OpenStack cloud. From this map, it's possible to
render a new config and upload it to the final destination.

In addition, this patch adds the ability to disable the
`Authentication` validation performed every time the provider is
instantiated. This way we'd be able to check the provider's configs
just once in an entire workflow execution. This also allows for testing
the new config data source without having to have an actual cloud in
the background. Lazy validation, FTW.

We could also remove validation entirely and let gophercloud error out
at the first interaction with the openstack cloud.