Networks v2: Network Data Source Fixes (#384)
* Networks v2: Network Data Source External Fix

PR #358 added support for querying networks by using the "external"
argument. However, this caused results to either contain only external
networks or only non-external networks.

When querying for an external network without the "external" argument,
no results would be found and Terraform would crash.

This fixes the behavior to only query with the "external" argument if
it is explicitly defined in the Terraform configuration.

* Networks v2: Network Data Source Crash Fix

The introduction of the "external" Gophercloud package extracted the
network results in a new way. This new method did not safely handle
results which did not contain any networks.

This commit resolves this issue.