Add tags support for Neutron networks
This allows us to set/get tags for networks via terraform templates,
and requires the recently added changes to gophercloud that enable
the relevant features of the Neutron APIs to be accessed.

Note the test is added in a new resource_openstack_networking_tags_v2_test.go
file, so that we can reduce duplication later in the series e.g
subnet/port/trunk resources can share the same template which should be faster.

Issue: #453