Make the tenantnetworks API not required (#485)
* Make the tenantnetworks API not required

Some OpenStack implementations (for instance, at my company) do not
allow the use of the os-tenant-networks API call. This causes a fatal
error when trying to read from an `openstack_compute_instance_v2`
resource. However, this error is not actually fatal as these properties
are not required for destruction. This PR changes the functionality of
the resource to reflect that.

Fixes #95

* Add an additional nil check to flattenInstanceNetworks

* Simplify code