Provider: Fix flavor detection (#551)
* Provider: Fix flavor detection

It appears that Terraform v0.12 won't use DefaultFunc when the field is
set to computed. While this might be a bug, we only have a few
occurrences of this in the provider, so I'm opting to work around this
instead of reporting it upstream and possibly have a legacy measure put
in place.

This commit will have the flavors still use the correct environment
variables. Detection is just deferred to outside the schema.

* Provider: fixing typo in error message