Networking v2: Fix Port admin_state_up behavior
If "admin_state_up" is not specified in the API request, then the port
should default to an admin_state_up of "UP".

Currently, this resource will always send a value of "false", even if
the user did not specify it in their resource configuration.

This commit fixes this behavior.