Networking v2: Fix boolean fields of network resource (#593)
openstack_networking_network_v2 was originally created with the fields
"admin_state_up" and "shared" as TypeString. This was to work around
Terraform mistaking an omitted boolean argument as "false".

Now that `GetOkExists` is available, we can correctly account for this
case.

This commit changes these two fields to TypeBool while still retaining
the prior behavior that TypeString was working around.