Networking v2: Deprecate openstack_networking_subnet_v2.host_routes (#668)
This commit deprecates the host_routes argument of the
openstack_networking_subnet_v2 resource. Given there is a schema issue
with setting the host_routes attribute and that using both this resource
and the openstack_networking_subnet_route_v2 resource will cause diffs
to occur, it's best to have the supported solution be to use the
dedicated resource.