Networking v2: Deprecate subnet allocation_pools for allocation_pool
This commit deprecates the openstack_networking_subnet_v2
allocation_pools argument in favor of allocation_pool.

allocation_pool is a singular word which makes more sense to use
declaratively.

In addition, allocation_pools was originally a TypeList, but the
OpenStack API will not respect the ordering defined by the user for
multiple pools, which triggered unneeded diffs.

allocation_pool is now a TypeSet which accounts for an unordered
response.