Networking v2: Make port dns_name computed
The OpenStack Compute service will set a port's dns_name to the name of
the instance when the OpenStack environment has the DNS extension
enabled. Because of this behavior, the dns_name needs to be set to
Computed or else the dns_name will be changed to an empty string on the
next apply.