compute_instance_v2: handle special case of 2 default networks
When you query for the details on a server in OpenStack, you get a list
of addresses assigned to that server which is keyed by pool. This
unfortunately causes problems becauses addresses are actually a JSON
object keyed by their network name. JSON objects are unordered by spec
and Go uses map[string]interface{} for JSON objects which is also
unordered. However this provider keeps the networks in a TypeList which
is order dependent. If you import a server then the networks will come
in a random order and terraform will insist it must rebuild the server.
Unfortunately there's no way to ensure that the ordering matches how it
came back from OpenStack due to this. This is only an issue when there
are two default networks such as for a provider with "public" and
"private" as default networks. Without a way to generically fix the
ordering of networks for imported servers, this fix just special cases
that situation to workaround the issue. fixes #1071

Signed-off-by: Doug Goldstein <doug.goldstein@rackspace.com>