Networking V2: fix 'binding.host_id' reset (#1082)
If the binding section is defined in the ressource definition (e.g.
just with "binding.vnic_type" set), but "binding.host_id" is left empty,
the port is created correctly and "binding.host_id" populated with the
id of the selected hypervisor. However, upon a subsequent plan/apply,
Terraform seems to reset "binding.host_id" to NULL. The port enters
status DOWN and connectivity is lost. Not sure this is the right fix,
but it seems the "host_id" should not be overriden in this case.

Signed-off-by: Gregory Vander Schueren <gregory.vanderschueren@tessares.net>