Allow to disable status check in resource_openstack_dns_zone_v2
This will allow terraform to tolerate some backend issues in Designate.
If zone is created in Designate, and even all the backends are down,
Designate will try to recover.
If only one of the backends is down, zone will be in ERROR state, but
will be served by other backends. In this case, waiting for ACTIVE state
also doesn't make sense.