Correctly delete openstack_networking_router_interface_v2 resources
When creating multiple openstack_networking_router_interface_v2
resources for different IPv6 subnets in the same network, openstack will
attach these to the same port on the router.

On destroy the current implementation will pass the port_id to
routers.RemoveInterface, which will trigger the deletion of the whole
port, including other openstack_networking_router_interface_v2 resources
on the same port.

This commit changes this behaviour (so that only the correct router interface
gets deleted) and adds a test case that ensures this does not break
again in the future.