Fix containerinfra cluster mergelabels behavior
When you add mergelabels and you actually add labels, the magnum API
will actually return you the complete set of labels under the "labels"
key. It means that if you want to add/override a couple of label while
setting merge_labels the creation will succeed but then terraform will
detect the rest of the labels as changed and will try to rebuid the
cluster.

This commit instead use the keys labels_{added,skipped,overriden}
when merge labels which contains the labels that the user has overriden.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>