Make container infra cluster aware of default worker nodegroup
The node_count returned in magnum for the cluster is the sum of all
nodegroup. So once you add a node_count, the correct value is actually
the node_count of the default worker node.

This commit take the node_count of this default worker node group in
priority and if there are any errors it takes the node_count of the
cluster instead.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>