Add containerinfra nodegroup support (#1364)
* Add containerinfra nodegroup resource

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Add containerinfra nodegroup data source

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Add containerinfra cluster zero node count support

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Add containerinfra nodegroup zero node count support

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Add containerinfra nodegroup resize node count support

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Fix containerinfra cluster mergelabels behavior

When you add mergelabels and you actually add labels, the magnum API
will actually return you the complete set of labels under the "labels"
key. It means that if you want to add/override a couple of label while
setting merge_labels the creation will succeed but then terraform will
detect the rest of the labels as changed and will try to rebuid the
cluster.

This commit instead use the keys labels_{added,skipped,overriden}
when merge labels which contains the labels that the user has overriden.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Add containerinfra nodegroup support for mergelabels

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Fix container infra nodegroup golangci-lint errors

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Fix container infra nodegroup image_id and flavor_id

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Fix containerinfra cluster template image/flavor_id

Previously the code set image_id and flavor_id which doesn't exist
instead of image/flavor. It would be way more meaningful to have argument
image_id and flavor_id instead but for backward compatibility let's not
do that for now.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Add OS_MAGNUM_IMAGE to containerinfra acceptance tests

As fedora atomic is no more available, we remove it from the code and
add OS_MAGNUM_IMAGE environment variable so that we can a supported
image in the CI later on.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Make container infra cluster aware of default worker nodegroup

The node_count returned in magnum for the cluster is the sum of all
nodegroup. So once you add a node_count, the correct value is actually
the node_count of the default worker node.

This commit take the node_count of this default worker node group in
priority and if there are any errors it takes the node_count of the
cluster instead.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>

* Fix container infra nodegroup acceptance tests

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>