Fixed markdown syntax
The upbound tool named "scrape" (used in the process of generating a crossplane provider from the terraform provider)
was failing when trying to convert markdown to JSON

In the case of `website/docs/r/compute_instance_v2.html.markdown`, this was the error:
example.hcl:5,21-22: Invalid expression; Expected the start of an expression, but found an invalid expression token.

In the case of `website/docs/r/networking_network_v2.html.markdown` the error was:
example.hcl:32,5-6: Invalid argument name; Argument names must not be quoted.