fix refresh for resource openstack_dns_recordset_v2 (#1581)
The resource was incorrectly failing to read in the actual record
values of the recordset resource when Terraform runs a refresh.  I
believe this change was made due to the designate system imposing its
own internal sorting on the record values instead of using the order
provided by the user when calling the API.  Due to this ordering,
and the "records" attribute being a list instead of a set, this would
cause the Terraform plan to be unnecessarily unstable until the user
reordered their record list to match whatever ordering that designate
wants to impose even though the order of the records is actually
unimportant.

This change converts the "records" attribute to a set since the
order of the records doesn't matter and record values should be distinct
within a given resource definition anyways. This attribute definition
change does not require a new schema version since a list of strings
and a set of strings are represented identically in the underlying
.tfstate file.

Also removed the support for stripping surrounding square brackets off
of IPv6 addresses as the functionality could actually prevent other
types of record values, such as TXT, being set correctly.