Document the secgroup_rule port-range exception
When you supply the full port-range 1:65535 neutron does something
"clever" and sets port_range_min and port_range_max to None. So when
the provider goes back to read out the port-range it comes out 0:0
instead of the expected 1:65535, and the rule gets replaced.

Every.
Single.
Time.

The solution of setting min and max to 0 is staring you right in the
face. But documenting this strange exception would be beneficial to
denser minds such as mine.