networking: add router flavor id option (#1991)
This adds an optional flavor_id argument to the openstack_networking_router_v2
resource, allowing users to specify a flavor for their routers.

To implement this, the router details are now read into
`RouterExtended` struct, which extends the `routers.Router` struct to include the
`FlavorID` field.

To test this feature the functional-networking workflow has been updated
to create an additional service provider. In addition, a new router
flavor and profile is created in the `stackenv.sh` script.

---------

Co-authored-by: Tobias Jungel <1773291+toanju@users.noreply.github.com>